newprojectmenu

+++ src/Main.java
@@ -600,8 +600,8 @@ public class Main {
 
                     JPanel newProjectPanel = new JPanel();
                     newProjectPanel.setBackground(new Color(50, 50, 50));
-                    int panelWidth = screenWidth/1.5;
-                    int panelHeight = screenHeight/1.5;
+                    double panelWidth = screenWidth/1.5;
+                    double panelHeight = screenHeight/1.5;
                     int x = (screenWidth - panelWidth) / 2;
                     int y = (screenHeight - panelHeight) / 2;
